feat: enhance user management interface and validation rules
- Updated UsersIndex component to improve mobile and desktop layouts, including responsive cards for user display.
- Added a new MultipleOf validation rule to ensure numeric inputs are multiples of a specified value.
- Migrated inventory and sale items quantity fields from INT to DECIMAL for better precision in stock management.
- Adjusted migration files to handle database-specific syntax for altering column types.

+++ app/Models/SaleItem.php
@@ -19,11 +19,11 @@ class SaleItem extends Model
     ];
 
     protected $casts = [
-        'quantity' => 'integer',
+        'quantity' => 'float',      // en vez de 'decimal:3'
         'unit_price' => 'decimal:2',
         'discount' => 'decimal:2',
         'total' => 'decimal:2',
-    ]; 
+    ];
 
     public function sale()
     {
